requirements.txt
openai
python-docx
loguru
pydantic
httpx
aiofiles
websockets
fastapi
schedule
beautifulsoup4
json_repair==0.*